--- data/sideItems.js
@@ -1,6 +1,0 @@
-const sideItems = [
-    { id:"dashboard_view", value:"Dashboard" },
-    { id:"users_view", value:"Users" },
-    { id:"products_view", value:"Products" },
-    { id:"admin_view", value:"Admin" },
-];
