+++ index.html
@@ -4,23 +4,17 @@
         <link rel="stylesheet" href="libs/webix/webix.css" type="text/css"> 
         <link rel="stylesheet" href="css/style.css" type="text/css"> 
         <script src="libs/webix/webix.js" type="text/javascript"></script>
-        <script src="data/sideItems.js" type="text/javascript"></script>
-        <script src="js/header.js" type="text/javascript"></script>
-        <script src="js/footer.js" type="text/javascript"></script>
-        <script src="js/main/side.js" type="text/javascript"></script>
-        <script src="js/main/dashboard/form.js" type="text/javascript"></script>
-        <script src="js/main/dashboard/table.js" type="text/javascript"></script>
-        <script src="js/main/users/toolbar.js" type="text/javascript"></script>
-        <script src="js/main/users/list.js" type="text/javascript"></script>
-        <script src="js/main/users/chart.js" type="text/javascript"></script>
-        <script src="js/main/products/treetable.js" type="text/javascript"></script>
-        <script src="js/main/main.js" type="text/javascript"></script>
     </head>
     <body>
-        <script>
+        <script type="module">
             
-            webix.ready(function(){
+            import {header, profilePopup} from "./js/header.js";
+            import {main} from "./js/main/main.js";
+            import {footer} from "./js/footer.js";
+            import {side} from "./js/main/side.js";
 
+            webix.ready(function(){
+                
                 webix.ui(
                     profilePopup
                 );
