+++ js/main/dashboard/table.js
@@ -1,4 +1,4 @@
-const dashboardTable = {
+export const dashboardTable = {
     id:"table_movies",
     view:"datatable",
     select:true,
